+++ src/strategies/iron_condor.rs
@@ -1092,7 +1092,7 @@ impl ProbabilityAnalysis for IronCondor {
             self.short_put.option.implied_volatility,
             self.long_call.option.implied_volatility,
             self.long_put.option.implied_volatility,
-        ]);
+        ])?;
 
         let mut profit_range = ProfitLossRange::new(
             Some(break_even_points[0]),
@@ -1125,7 +1125,7 @@ impl ProbabilityAnalysis for IronCondor {
             self.short_put.option.implied_volatility,
             self.long_call.option.implied_volatility,
             self.long_put.option.implied_volatility,
-        ]);
+        ])?;
 
         let mut loss_range_lower =
             ProfitLossRange::new(None, Some(break_even_points[0]), Positive::ZERO)?;